feat: divide collection view data source from TagCategoryViewController
#14
